backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

sqlmodel==0.0.16
sqlalchemy==2.0.29
psycopg2-binary==2.9.9  # Or psycopg2 if using non-binary

pydantic==2.6.4
python-dotenv==1.0.1

pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
scipy==1.13.0

mlflow==2.12.1
xgboost==2.0.3 
joblib==1.4.2
pymysql==1.1.0
